requirements.txt
svg.path==4.0.2
tinycss2==1.1.0
